final-project - 프로필 커스터마이징하고 디자인만 바꾸면 됨/backend/saving/serializers.py
from rest_framework import serializers
from .models import SavingProducts, SavingOptions


class SavingProductsSerializer(serializers.ModelSerializer):
    class Meta:
        model = SavingProducts
        fields = "__all__"


class SavingOptionsSerializer(serializers.ModelSerializer):
    class Meta:
        model = SavingOptions
        fields = "__all__"
